feat: add `AddOrleansServerRuntime` to `HostApplicationBuilder`, implement clustering, event sourcing, hosting, and activity propagation defaults
