+++ src/Hosting/HostApplicationBuilderExtensions.Orleans.cs
@@ -14,6 +14,9 @@ using Orleans.Configuration;
 /// </summary>
 public static partial class HostApplicationBuilderExtensions
 {
+    private const string OrleansApplicationSourceName = "Microsoft.Orleans.Application";
+    private const string OrleansApplicationMeterName = "Microsoft.Orleans";
+
     /// <summary>
     /// Configures the host application builder to use a runtime server.
     /// This method extends the HostApplicationBuilder with functionality
@@ -25,53 +28,20 @@ public static partial class HostApplicationBuilderExtensions
     /// <returns>
     /// The configured <see cref="HostApplicationBuilder"/> instance.
     /// </returns>
-    public static HostApplicationBuilder UseRuntimeServer(this HostApplicationBuilder builder)
+    public static HostApplicationBuilder AddOrleansServerRuntime(this HostApplicationBuilder builder)
     {
         ArgumentNullException.ThrowIfNull(builder);
+
+        // Setup Orleans Defaults
         return builder
+            .AddClusteringDefaults()
             .AddEventSourcingDefaults()
             .AddHostingDefaults()
-            .AddActivityPropagation();
-    }
-
-    private static HostApplicationBuilder AddEventSourcingDefaults(this HostApplicationBuilder builder)
-    {
-        ArgumentNullException.ThrowIfNull(builder);
-        return builder
-            .UseOrleans(siloBuilder => siloBuilder
-                .AddLogStorageBasedLogConsistencyProviderAsDefault()
-                .AddStateStorageBasedLogConsistencyProviderAsDefault());
-    }
-
-    private static HostApplicationBuilder AddHostingDefaults(this HostApplicationBuilder builder)
-    {
-        ArgumentNullException.ThrowIfNull(builder);
-        builder
-            .Services
-            .Configure<ClusterOptions>(configureOptions => builder
-                .Configuration
-                .Bind(configureOptions));
-
-        builder
-            .UseOrleans(siloBuilder =>
-            {
-                var hostingMode = builder
-                    .Configuration
-                    .GetValue<string>("Hosting");
-
-                _ = hostingMode switch
-                {
-                    "Kubernetes" => siloBuilder.UseKubernetesHosting(),
-                    _ => siloBuilder,
-                };
-            });
-        return builder;
-    }
-
-    private static HostApplicationBuilder AddActivityPropagation(this HostApplicationBuilder builder)
-    {
-        return builder
-            .UseOrleans(siloBuilder => siloBuilder
-                .AddActivityPropagation());
+            .AddActivityPropagation()
+            .AddServiceDefaults(
+                tracingProvider => tracingProvider
+                    .AddSource(OrleansApplicationSourceName),
+                metersProviderBuilder: metricsProviderBuilder => metricsProviderBuilder
+                    .AddMeter(OrleansApplicationMeterName));
     }
 }
